+++ core/src/OptParser.cpp
@@ -49,7 +49,7 @@ namespace {
    * @param y        Return value y
    * @param usage    String containing some usage information which is printed when there is an error.
    */
-  void parsePosition(TString parseMe, float& x, float& y, TString usage) {
+  void parsePosition(TString parseMe, double& x, double& y, TString usage) {
     if (parseMe == TString("default")) {
       x = -1.;
       y = -1.;
@@ -144,7 +144,7 @@ namespace {
    * @param min     Return value
    * @param max     Return value
    */
-  bool parseRange(TString parseMe, float& min, float& max) {
+  bool parseRange(TString parseMe, double& min, double& max) {
     if (parseMe == TString("default")) {
       min = -104;
       max = -104;
@@ -185,7 +185,7 @@ namespace {
    * @param name    Return string
    * @param value   Return value
    */
-  bool parseAssignment(TString parseMe, TString& name, float& value) {
+  bool parseAssignment(TString parseMe, TString& name, double& value) {
     TString valueStr;
     parseAssignment(parseMe, name, valueStr);
     value = valueStr.Atof();
@@ -536,9 +536,9 @@ void OptParser::parseArguments(int argc, char* argv[]) {
                                              "of the y variable to a given range. "
                                              "Format: --scanrangey min:max.",
                                              false, "default", "string");
-  TCLAP::ValueArg<float> scaleerrArg("", "scaleerr", "Scale the errors by this number", false, -999., "float");
-  TCLAP::ValueArg<float> scalestaterrArg("", "scalestaterr", "Scale the STAT only errors by this number", false, -999.,
-                                         "float");
+  TCLAP::ValueArg<double> scaleerrArg("", "scaleerr", "Scale the errors by this number", false, -999., "double");
+  TCLAP::ValueArg<double> scalestaterrArg("", "scalestaterr", "Scale the STAT only errors by this number", false, -999.,
+                                          "double");
   TCLAP::ValueArg<std::string> plotoriginArg(
       "", "origin", "Plot Origin on 2D plots. Default 0:0. Can move to another location. Format: --origin min:max",
       false, "default", "string");
@@ -640,10 +640,10 @@ void OptParser::parseArguments(int argc, char* argv[]) {
       "Format: --grouppos xmin:ymin in normalized coordinates [0,1]. To use default values "
       "for one coordinate, use 'def': --grouppos def:y.",
       false, "default", "string");
-  TCLAP::ValueArg<float> printSolXArg("", "printsolx", "x coordinate to print solution at in 1D plots", false, -999.,
-                                      "float");
-  TCLAP::ValueArg<float> printSolYArg("", "printsoly", "y coordinate to shift solution by in 1D plots", false, -999.,
-                                      "float");
+  TCLAP::ValueArg<double> printSolXArg("", "printsolx", "x coordinate to print solution at in 1D plots", false, -999.,
+                                       "double");
+  TCLAP::ValueArg<double> printSolYArg("", "printsoly", "y coordinate to shift solution by in 1D plots", false, -999.,
+                                       "double");
   /// TCLAP::ValueArg<std::string> queueArg("q","queue","Batch queue to submit to. If none is given then the scripts
   /// will be written but not submitted.", false, "", "string");
   TCLAP::ValueArg<int> batchstartnArg(
@@ -812,13 +812,13 @@ void OptParser::parseArguments(int argc, char* argv[]) {
                                 "ID of color to be used for the combination. "
                                 "Default: 0 for first scanner, 1 for second, etc.",
                                 false, "int");
-  TCLAP::MultiArg<float> CLArg(
+  TCLAP::MultiArg<double> CLArg(
       "", "CL",
       "Confidence Levels to be computed and plotted in percent. This argument can be passed multiple times.\n"
       "Default will print 1 & 2 (3) sigma confidence levels\n"
       "Syntax: --CL 90  \n"
       "alternative: -cl 95.45  \n",
-      false, "float");
+      false, "double");
 
   TCLAP::MultiArg<int> clsArg(
       "", "cls",
@@ -969,13 +969,13 @@ void OptParser::parseArguments(int argc, char* argv[]) {
       "To set just the start values in the second combination, do\n"
       "Example: --start none --start 'g=1.7,r_dk=0.09' \n",
       false, "string");
-  TCLAP::MultiArg<float> snArg("", "sn",
-                               "--sn x. Save nuisances to parameter cache file at certain points after a "
-                               "1d scan was performed. This can be used to set these as starting points "
-                               "for further scans. "
-                               "Parameters will be saved for scan point bin that contains x. "
-                               "Angles have to be given in radians.",
-                               false, "float");
+  TCLAP::MultiArg<double> snArg("", "sn",
+                                "--sn x. Save nuisances to parameter cache file at certain points after a "
+                                "1d scan was performed. This can be used to set these as starting points "
+                                "for further scans. "
+                                "Parameters will be saved for scan point bin that contains x. "
+                                "Angles have to be given in radians.",
+                                false, "double");
   TCLAP::MultiArg<std::string> sn2dArg("", "sn2d",
                                        "Save nuisances as for --sn (1d case) but for 2d scans at the given points. "
                                        "Format: x:y",
